TriggerDatabase class established .
