+++ MissionEditor/MissionEditor.vcxproj.filters
@@ -315,6 +315,9 @@
     <ClCompile Include="TriggerDef.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="TriggerDatabase.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="res\VersionInfo.rc">
@@ -656,6 +659,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="TriggerDef.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="TriggerDatabase.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
